refactor: implement Pydantic schema refactoring - Add base schema classes and configuration - Migrate shared kernel value objects - Update tests with descriptive names and edge cases - Remove deprecated schemas and files

+++ services/app/src/contexts/seedwork/shared/endpoints/decorators/timeout_after.py
@@ -3,7 +3,7 @@ import functools
 from anyio import fail_after
 
 
-def timeout_after(timeout: int = 10):
+def timeout_after(timeout: float = 10):
     def decorator(func):
         @functools.wraps(func)
         async def wrapper(*args, **kwargs):
